feat: cleanup payments module
